refactor(server): extract tests from library module
